pyproject.toml
[project]
name = "media-bridge"
version = "0.1.0"
description = "MQTT bridge for living room media system - TV/audio state to Home Assistant"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
audio = [
    "dbus-python>=1.3.2",
]

[project.scripts]
media-bridge = "media_bridge.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/media_bridge"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
